Added authentication middleware
An authentication middleware has been added to check for token validation. If the token has expired, it will refresh the token. There's also a function that returns the header with the JWT token attached to it. A query has been added to search in the Spotify Web API for albums, tracks and artists based on a specified query.